move some files out of folders that do not need tests right now
