+++ src/Components/TileBoard/InputValidation/Form.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react'
 
-import { useFormContext } from '../FormContext'
+import { useFormContext } from './FormProvider/FormContext'
 
 type FormProps = {
     onSubmit?: (arg: any) => void
